fix: dupes for sonarcloud
