+++ sonar-project.properties
@@ -1,0 +1,5 @@
+sonar.projectKey=comps
+sonar.sources=.
+sonar.python.version=3.13
+sonar.exclusions=static/**,templates/**,uploads/**,node_modules/**,**/__pycache__/**
+sonar.cpd.exclusions=migrations/**,templates/**,static/**
